moved utmtolatlng function
